Polish rain drop watershed
